+++ examples/resources/awscc_amazonmq_configuration/import.sh
@@ -1,0 +1,1 @@
+$ terraform import awscc_amazonmq_configuration.example "id"