+++ internal/aws/wisdom/knowledge_base_singular_data_source_gen.go
@@ -60,7 +60,8 @@ func knowledgeBaseDataSource(ctx context.Context) (datasource.DataSource, error)
 		//	{
 		//	  "enum": [
 		//	    "EXTERNAL",
-		//	    "CUSTOM"
+		//	    "CUSTOM",
+		//	    "MESSAGE_TEMPLATES"
 		//	  ],
 		//	  "type": "string"
 		//	}